requirements.txt
Pillow == 10.4.0